Base gameplay: turns, action points, movement

+++ core/src/main/java/com/dai/ai/ITraversable.java
@@ -8,6 +8,6 @@ public interface ITraversable {
 
     public boolean isTraversable();
 
-    public float getCostModifier();
+    public int getCostModifier();
 
 }
